+++ osu.Game/Beatmaps/Formats/LegacyBeatmapDecoder.cs
@@ -332,7 +332,7 @@ namespace osu.Game.Beatmaps.Formats
                     break;
 
                 case @"GridSize":
-                    beatmap.GridSize = Parsing.ParseFloat(pair.Value);
+                    beatmap.GridSize = Parsing.ParseInt(pair.Value);
                     break;
 
                 case @"GridType":
